Feat: generate grid display
