+++ src/Style/NavBar.css
@@ -9,7 +9,7 @@
 }
 
 .hamburger-icon {
-  display: none; /* Oculta el botón de hamburguesa por defecto */
+  display: none;
 }
 
 .navbar img {
