+++ src/Style/myReservation.css
@@ -23,8 +23,7 @@
 }
 
 .card-reserv-cont {
-  /* background-color: bisque; */
-  width: 95vw;
+  width: 95%;
   padding: 10px;
 }
